+++ frontend/src/features/profile/ProfileKeywordView.tsx
@@ -1,0 +1,3 @@
+export const ProfileKeywordView = () => {
+  return <div>ProfileKeywordView</div>;
+};
